remove advanced manager
