--- octobot_commons/tentacles_management/config_manager.py
@@ -1,34 +1,0 @@
-#  Drakkar-Software OctoBot-Commons
-#  Copyright (c) Drakkar-Software, All rights reserved.
-#
-#  This library is free software; you can redistribute it and/or
-#  modify it under the terms of the GNU Lesser General Public
-#  License as published by the Free Software Foundation; either
-#  version 3.0 of the License, or (at your option) any later version.
-#
-#  This library is distributed in the hope that it will be useful,
-#  but WITHOUT ANY WARRANTY; without even the implied warranty of
-#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-#  Lesser General Public License for more details.
-#
-#  You should have received a copy of the GNU Lesser General Public
-#  License along with this library.
-from octobot_commons.config import load_config
-
-
-def reload_tentacle_config(
-    config, config_key, tentacle_config_path, raised_exception=None
-):
-    """
-    Reload tentacle configuration
-    :param config: the current config
-    :param config_key: the config key to reload
-    :param tentacle_config_path: the tentacle config path
-    :param raised_exception: the exception to be raised
-    :return: the reloaded config
-    """
-    config[config_key] = load_config(tentacle_config_path, False)
-    if config[config_key] is None:
-        raise raised_exception if raised_exception is not None else Exception
-
-    return config
